@@ -2,13 +2,15 @@
 
 import resolve from '@rollup/plugin-node-resolve'
 import commonjs from '@rollup/plugin-commonjs'
+import typescript from '@rollup/plugin-typescript';
 import pkg from './package.json'
 
 export default {
-	input: `src/index.js`,
+	input: `src/index.ts`,
 	plugins: [
-		commonjs(),
 		resolve(),
+		commonjs(),
+		typescript()
 	],
 	output: [
 		{
